refactor: rename `FilterPhotoCard` to `filter-photo-card`
Renamed related .stories .module.scss and other files.
