+++ features/filter-photo-card/filter-photo-card.module.scss
