+++ features/filter-photo-card/index.ts
@@ -1,0 +1,1 @@
+export * from './filter-photo-card';
