+++ features/index.ts
@@ -2,3 +2,4 @@ export * from './DEPRECATED_Modal';
 export * from './photo-slider';
 export * from './photo-gallery';
 export * from './add-photo-card';
+export * from './filter-photo-card';
